Added test for clear cell
