+++ core/test/java/board/MapBoardTests.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static utility.BoardUtils.fillBoardWithWhiteStones;
+import static board.providers.BoardProviders.fillBoardWithWhiteStones;
 
 class MapBoardTests {
