+++ core/test/java/game/FreedomLineTests.java
@@ -1,5 +1,6 @@
 package game;
 
+import board.providers.BoardProviders;
 import it.units.sdm.project.board.Position;
 import it.units.sdm.project.board.Piece;
 import it.units.sdm.project.game.FreedomLine;
@@ -7,7 +8,6 @@ import it.units.sdm.project.board.Board;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import utility.BoardUtils;
 
 import java.util.SortedSet;
 
@@ -22,7 +22,7 @@ class FreedomLineTests {
             + " 2 B  W  W  W  W  W  W  W\n"
             + " 1 B  W  W  B  W  W  W  W\n"
             + "   A  B  C  D  E  F  G  H";
-    private final Board<Piece> board = BoardUtils.parseBoardFromString(printedBoard, 8, 8);
+    private final Board<Piece> board = BoardProviders.parseBoardFromString(printedBoard, 8, 8);
 
     @ParameterizedTest
     @MethodSource("game.providers.FreedomLineProviders#provideInitialLinePositionFor8x8BoardWithException")
