+++ core/test/java/game/FreedomPointsCounterTests.java
@@ -1,5 +1,6 @@
 package game;
 
+import board.providers.BoardProviders;
 import com.badlogic.gdx.graphics.Color;
 import it.units.sdm.project.game.FreedomPointsCounter;
 import it.units.sdm.project.board.Piece;
@@ -8,7 +9,6 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import utility.BoardUtils;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -26,7 +26,7 @@ class FreedomPointsCounterTests {
                         + " 2 -  -  -  -  -  -  -  -\n"
                         + " 1 -  -  -  -  -  -  -  -\n"
                         + "   A  B  C  D  E  F  G  H";
-        Board<Piece> board = BoardUtils.parseBoardFromString(printedBoard, 8, 8);
+        Board<Piece> board = BoardProviders.parseBoardFromString(printedBoard, 8, 8);
         assertEquals(printedBoard, board.toString());
     }
 
@@ -34,7 +34,7 @@ class FreedomPointsCounterTests {
     @ParameterizedTest
     @MethodSource("game.providers.FreedomPointsCounterProviders#printedBoardsProvider")
     void testGetWinner(String printedBoard, int numberOfRows, int numberOfColumns, int blackScore, int whiteScore) {
-        Board<Piece> board = BoardUtils.parseBoardFromString(printedBoard, numberOfRows, numberOfColumns);
+        Board<Piece> board = BoardProviders.parseBoardFromString(printedBoard, numberOfRows, numberOfColumns);
         FreedomPointsCounter freedomPointsCounter = new FreedomPointsCounter(board);
         Assertions.assertEquals(freedomPointsCounter.getPlayerScore(Color.BLACK), blackScore);
         Assertions.assertEquals(freedomPointsCounter.getPlayerScore(Color.WHITE), whiteScore);
